+++ src/EasyGoal.Backend.WebApi/Contracts/Responses/DecisionHelper/DecisionHelperCriterionResponse.cs
@@ -2,6 +2,6 @@
 
 namespace EasyGoal.Backend.WebApi.Contracts.Responses.DecisionHelper;
 
-public sealed record DecisionHelperCriterionResponse(Guid Id, string Name, decimal Weight, DecisionHelperCriterionType Type)
+public sealed record DecisionHelperCriterionResponse(Guid Id, string Name, double Weight, DecisionHelperCriterionType Type)
 {
 }
